requirements.txt
boto3
matplotlib
numpy
torch
torchvision
pillow